refactor: renomeacao e formatacao

+++ Loteria/src/dominio/Premio.java
@@ -1,11 +1,12 @@
 package dominio;
 
 public class Premio {
+    //Classe responsavel pelo calculo do premio
     private long premioTotal;
 
     public Premio(){}
 
-    public void calculaPremioTotal(int valorAposta, int quantidadeApostas){
+    public void calculaPremioTotal(double valorAposta, int quantidadeApostas){
         this.premioTotal += valorAposta * quantidadeApostas;
     }
